Implement JSON-API expectations in JsonApiSerializer
Based on the format of JSON-API v1.0, I had to include one more hook from the
Scope in order to inject the `relationships:{}` links into the data.

It should be backwards-compatible, because there is a standard implementation
that does nothing in `SerializerAbstract` (which can then be overridden by the
concrete Serializer implementation that needs it).