Add support for nested includes
It was necessary to create one more hook from Scope to the serializer that
gives the serializer the ability to modify the final list of sideloaded includes
before it is merged into the data.

The reason is the JSON API spec. It is forbidden to include the root object into
the `included` array. But this would happen if any other object included relates
to the root object (i.e. references it).