Omit default or requested available includes
The new `parseExcludes()` method accepts a string or array similar to
`parseIncludes()` without support for modifiers.

Paths found in the list of parsed excludes will be omitted from
transformations. This allows `$defaultIncludes` to be excluded
from a transformation at the Manager level instead of needing
to subclass the Transformer and override/clear the
`$defaultIncludes` property.