Get JSON API type from transformer instead of $resoruceKey
BREAKING: While technically serializing type as an attribute is a bug, the previous implementation gave no warning or error if type was included. After this the type will be striped out of attributes.

BREAKING: `type` is now a required element returned from the transformer (similar to id).

BREAKING: The $resourceKey will be ignored now, and the type used for serializing.