Allow use of newer PHPUnit on newer versions of PHP
This allows the 0.x series of this library to have passing tests on PHP
5.4 all the way up to PHP 7.3. I added an allowance for using the
PHPUnit 7 series on versions of PHP that support it, and I made a few
slight changes to tests and phpunit.xml.dist that allow the tests to
continue running on both the older version of PHPUnit and the newer
versions.

FYI, when you choose to upgrade to PHPUnit 8.x for the 1.x series of this
library, you'll need to make more changes to the tests that won't be
backwards compatible with earlier versions of PHPUnit.